pyproject.toml
[project]
name = "nonebot-plugin-ehentai"
version = "0.1.2"
description = "下载eh本子并发送"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.11.18",
    "img2pdf>=0.6.1,<1.0.0",
    "nonebot-plugin-alconna>=0.54.2,<1.0.0",
    "nonebot-plugin-apscheduler>=0.5.0,<1.0.0",
    "nonebot-plugin-localstore>=0.7.4,<1.0.0",
    "nonebot2>=2.4.1,<3.0.0",
    "pypdf>=5.4.0",
]

[project.urls]
homepage = "https://github.com/MaxCrazy1101/nonebot-plugin-ehentai"
repository = "https://github.com/MaxCrazy1101/nonebot-plugin-ehentai"

[dependency-groups]
dev = [
    "nonebot2[fastapi]>=2.4.2,<3.0.0",
    "nonebot-adapter-onebot>=2.4.6,<3.0.0",
    "nonebug>=0.3.7,<1.0.0",
    "pytest-xdist>=3.6.1,<4.0.0",
    "pytest-asyncio>=0.26.0,<1.0.0",
]

[tool.uv]
default-groups = ["dev"]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.nonebot]
plugins = ["nonebot_plugin_ehentai"]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" },
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
pythonpath = ["."]
addopts = [
    "-v",               # 详细输出
    "-s",               # 显示打印信息
    "--tb=short",       # 简短的错误回溯
    "-ra",              # 显示所有测试结果摘要
    "--strict-markers", # 严格标记模式
    # "--doctest-modules", # 运行文档测试
    "--import-mode=prepend", # 导入模式
]

[tool.isort]
profile = "black"
line_length = 90
length_sort = true
skip_gitignore = true
force_sort_within_sections = true
extra_standard_library = ["typing_extensions"]

[tool.ruff]
line-length = 90
target-version = "py310"

[tool.ruff.lint]
select = ["E", "W", "F", "UP", "C", "T", "PYI", "PT", "Q"]
ignore = ["E402", "C901"]

[tool.pyright]
pythonVersion = "3.10"
pythonPlatform = "All"
defineConstant = { PYDANTIC_V2 = true }
typeCheckingMode = "standard"
reportShadowedImports = false
disableBytesTypePromotions = true
